+++ ccxt.php
@@ -42,7 +42,7 @@ class DDoSProtection       extends NetworkError {}
 class RequestTimeout       extends NetworkError {}
 class ExchangeNotAvailable extends NetworkError {}
 
-$version = '1.4.12';
+$version = '1.4.22';
 
 $curl_errors = array (
     0 => 'CURLE_OK',